% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{CheckSearchInput}
\alias{CheckSearchInput}
\title{CheckSearchInput}
\usage{
CheckSearchInput(input, output, session)
}
\value{
TRUE / FALSE
}
\description{
CheckSearchInput figures out if there is any input in `search_input` at the moment. 
Don't forget to call from appropriate namespace
}
