% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{csvFile}
\alias{csvFile}
\title{csvFile}
\usage{
csvFile(input, output, session, stringsAsFactors, file_name = "file")
}
\arguments{
\item{file_name}{is the name of the file; needs to match \code{file_name} fed into \code{csvFileInput}, the UI function.}

\item{stringAsFactors}{is input for \code{read.csv}}
}
\value{
dataframe with contents of the CSV file
}
\description{
csvFile processes the input from csvFileInput and reads it into a \code{data.frame}. Adapted from \link{https://shiny.rstudio.com/articles/modules.html}
}
