% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{csvFile_local}
\alias{csvFile_local}
\title{csvFile_local}
\usage{
csvFile_local(input, output, session, directory, file_name)
}
\arguments{
\item{directory}{name of the directory to look for a file}

\item{file_name}{name of the csv file contained in dir}
}
\value{
a `data.frame` of the content of the file_name
}
\description{
csvFile_local is an alternative to `csvFile()`, 
this reads a local file instead of calling the upload function
calls on `read.csv()`
}
\details{
here be details
}
