% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{csvFileInput}
\alias{csvFileInput}
\title{csvFileInput}
\usage{
csvFileInput(name_space, file_name = "file", label = "Input file")
}
\arguments{
\item{name_space}{The name space within this is located}

\item{file_name}{Variable, to be called as \code{input[[file_name]]}}

\item{label}{Is the label displayed to user}
}
\value{
a \code{input$file_name}: from doc: [Returned] dataframe contains one row for each selected file, and the following columns [...] datapath, name, size
}
\description{
UI function that lets user input a CSV to be read into \code{data.frame} by server function csvFile. Adapted from \link{https://shiny.rstudio.com/articles/modules.html}
}
