% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{deliveryConverter}
\alias{deliveryConverter}
\title{deliveryConverter}
\usage{
deliveryConverter(input_data_frame)
}
\arguments{
\item{input_data_frame}{is a \code{data.frame} that includes at least sentiment scores and relevance columns for individual delivery themes. E.g. read in with \code{csvFile()} in this package.}
}
\value{
time series object \code{xts} formatted, containing sentiment scores by week.
}
\description{
deliveryConverter takes a data_frame with sentiment scores and converts it to a time series object (\code{xts / zoo}) to feed into a dygraph.
It's based on Lars' initial process chain.
}
\details{
processes the input from csvFileInput and reads it into a \code{data.frame}. Adapted from \link{https://shiny.rstudio.com/articles/modules.html}

the exception is modeled after \link{https://shiny.rstudio.com/gallery/file-upload.html} FIX THE CAPS IN THE TITLE
}
