% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{GlobToRegex}
\alias{GlobToRegex}
\title{GlobToRegex}
\usage{
GlobToRegex(input_string)
}
\arguments{
\item{input_string:}{string, optionally containing wildcards such as `*` or `?`}
}
\value{
string to be used as input into \code{grep}
}
\description{
GlobToRegex takes a string as input_text and converts it to a regex to be used by \code{grep}.
}
\examples{
`run*` --> `run.*`; `r?n` --> `r.{1}n`
}
