% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{MeanSentimentPerDay}
\alias{MeanSentimentPerDay}
\title{MeanSentimentPerDay}
\usage{
MeanSentimentPerDay(data_frame, sentiment_column, date_column, days_data_frame,
  origin_date)
}
\arguments{
\item{data_frame:}{the data.frame to be processed, where}

\item{sentiment_column:}{a column of data_frame that contains sentiment scores}

\item{date_column:}{a column of data_frame that contains the date of the object}

\item{days_data_frame:}{a data.frame of days to be matched to the days in date_column}

\item{origin_date:}{beginning date, CHECK THE FORMAT ON THIS ONE}
}
\value{
A time series  (\code{"xts" "zoo"}, created by \code{as.xts}) of mean sentiment scores, format: index=date, mean_sentiment. Note that it will return an empty \code{xts} object if the input has zero rows.
}
\description{
MeanSentimentPerDay computes mean sentiment score by day.
}
