% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{ResultChecker}
\alias{ResultChecker}
\title{ResultChecker}
\usage{
ResultChecker(in_data, min_rows = 1)
}
\arguments{
\item{in_data}{a \code{data.frame}, to be checked}

\item{min_rows}{minumum number of rows a in_data needs to have to pass}
}
\value{
boolean: Passed TRUE or FALSE
}
\description{
ResultChecker runs sanity tests on data.frames, e.g. produced by search
}
