% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{SearchTerm}
\alias{SearchTerm}
\title{SearchTerm}
\usage{
SearchTerm(input, output, session, remove_whitespace = FALSE)
}
\arguments{
\item{input}{string from a \code{textInput} search box as \code{input[['search_term']]}}

\item{remove_whitespace}{if set to \code{TRUE}, remove leading and trailing whitespace}

\item{input}{event from an \code{ActionButton} button as \code{execute_search}}
}
\value{
reactive expression object created by \code{eventReactive}
}
\description{
SearchTerm is the server-side function of the SearchTerm module, cf \link{SearchTermInput}
It takes the input from a search box after "Search" action button has been clicked.
}
\details{
\code{eventReactive} docs: \url{http://shiny.rstudio.com/reference/shiny/latest/observeEvent.html}
}
