% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{SearchTermInput}
\alias{SearchTermInput}
\title{SearchTermInput}
\usage{
SearchTermInput(name_space, label = "Search box", button_label = "Search",
  help_text = NULL)
}
\arguments{
\item{name_space}{the associated namespace, what the docs call \code{id}}

\item{label}{a string to be supplied to \code{label = } of \code{textInput}}

\item{button_label}{a string to be supplied to \code{label = } of \code{ActionButton}}

\item{help_text}{a string to appear as help text (mouse over) on \code{textInput}. Defaults to \code{NULL}}
}
\value{
\code{input[['search_term']]} from \code{textInput} and \code{input[['execute_search']]} from \code{ActionButton}
}
\description{
SearchTermInput is the UI-side function of the SearchTerm module, cf \link{SearchTerm}
It creates a tagList to be fed into any \code{sidebarPanel} 
Taglist contains a box for text input and an action button.
}
\details{
The \code{div} part is from SO, \url{https://stackoverflow.com/questions/16449252/tooltip-on-shiny-r#16451601}
}
