% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{TextParser}
\alias{TextParser}
\title{TextParser}
\usage{
TextParser(input_text, lower_case = TRUE, remove_stopwords = FALSE,
  remove_numbers = FALSE)
}
\arguments{
\item{input_text}{a string}

\item{lower_case}{if set to \code{TRUE}, will lowercase the \code{input_text}}

\item{remove_stopwords}{if set to \code{TRUE}, will remove all stopwords (currently the ones from pckg tokenizers)}

\item{remove_numbers}{CURRENTLY NOT IMPLEMENTED}
}
\value{
a list with one entry, a vector of tokens
}
\description{
The TextParser takes a string of words and tokenizes it into a list
}
\details{
note that this deletes all punctuation by default
}
