% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{TimeseriesConverter}
\alias{TimeseriesConverter}
\title{The TimeseriesConverter takes an input_data_frame and converts it into a \code{xts} frame to be plotted in a dygraph.}
\usage{
TimeseriesConverter(input_data_frame, col_name)
}
\arguments{
\item{input_data_frame}{is a \code{data.frame} that includes at least sentiment scores and relevance columns for individual delivery themes. E.g. read in with \code{csvFile()} in this package.}

\item{col_name}{is a string giving a title to the result-column of the output time series. This will e.g. show up as label in a dygraph.}
}
\value{
time series object \code{xts} formatted, containing sentiment scores by week.
}
\description{
It is a generalized version of the deliveryConverter, but e.g. controls for zero length input.
}
