% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{WordCounter}
\alias{WordCounter}
\title{WordCounter}
\usage{
WordCounter(list_of_tokens)
}
\arguments{
\item{list_of_tokens}{a list with one entry containing a tokenized string (vector)}
}
\value{
a vector mapping tokens to counts
}
\description{
The WordCounter counts how often each token in list_of_tokens occurs.
}
\details{
as of now, the counts are only accurate if there is only one item in the \code{list_of_tokens}. 
It will work well with output from \link{TextParser}.
}
