% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimenttools.R
\name{wordfinder}
\alias{wordfinder}
\title{Wordfinder}
\usage{
wordfinder(input_text, search_term, lower_case = TRUE)
}
\arguments{
\item{input_text:}{string}

\item{search_term:}{string that will be checked for existence in input_text}

\item{lower_case:}{as of now, argument is not used; all text is lowercased}
}
\value{
TRUE / FALSE
}
\description{
Wordfinder takes a string as input_text and returns TRUE if search_term is contained.
}
\examples{
wordfinder("This is a sentence", "sentence")
NB THIS WILL MESS WITH EMOTICONS, NO?
}
